move tokenstorage implementations into own packages
