+++ oauth/token_upload_test.go
@@ -21,6 +21,7 @@ import (
 	"github.com/redhat-appstudio/service-provider-integration-operator/api/v1beta1"
 	api "github.com/redhat-appstudio/service-provider-integration-operator/api/v1beta1"
 	"github.com/redhat-appstudio/service-provider-integration-operator/pkg/spi-shared/tokenstorage"
+	"github.com/redhat-appstudio/service-provider-integration-operator/pkg/spi-shared/tokenstorage/vaultstorage"
 	"github.com/stretchr/testify/assert"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	"k8s.io/apimachinery/pkg/runtime"
@@ -45,7 +46,7 @@ func TestTokenUploader_ShouldUploadWithNoError(t *testing.T) {
 
 	strg := tokenstorage.NotifyingTokenStorage{
 		Client: cl,
-		TokenStorage: tokenstorage.TestTokenStorage{
+		TokenStorage: vaultstorage.TestTokenStorage{
 			StoreImpl: func(ctx context.Context, token *v1beta1.SPIAccessToken, data *v1beta1.Token) error {
 				assert.Equal(t, cntx, ctx)
 				assert.Equal(t, "jdoe", data.Username)
@@ -88,7 +89,7 @@ func TestTokenUploader_ShouldFailTokenNotFound(t *testing.T) {
 
 	strg := tokenstorage.NotifyingTokenStorage{
 		Client: cl,
-		TokenStorage: tokenstorage.TestTokenStorage{
+		TokenStorage: vaultstorage.TestTokenStorage{
 			StoreImpl: func(ctx context.Context, token *v1beta1.SPIAccessToken, data *v1beta1.Token) error {
 				assert.Fail(t, "This line should not be reached")
 				return nil
@@ -126,7 +127,7 @@ func TestTokenUploader_ShouldFailOnStorage(t *testing.T) {
 
 	strg := tokenstorage.NotifyingTokenStorage{
 		Client: cl,
-		TokenStorage: tokenstorage.TestTokenStorage{
+		TokenStorage: vaultstorage.TestTokenStorage{
 			StoreImpl: func(ctx context.Context, token *v1beta1.SPIAccessToken, data *v1beta1.Token) error {
 				return fmt.Errorf("storage disconnected")
 			},
