+++ pkg/spi-shared/tokenstorage/vaultstorage/tokenstorage_vault_test.go
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package tokenstorage
+package vaultstorage
 
 import (
 	"context"
