+++ pkg/spi-shared/tokenstorage/vaultstorage/vault_auth.go
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package tokenstorage
+package vaultstorage
 
 import (
 	"errors"
