+++ pkg/spi-shared/tokenstorage/vaultstorage/vault_auth_test.go
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package tokenstorage
+package vaultstorage
 
 import (
 	"io/fs"
